Improve registration page functionality
Now correctly displays an error when a user try to use an already registered mail. No more refresh page on passwords not matching error and for now redirects on homepage when registration is successful. It should auto-login user and set cookie, maybe redirecting to a special page before returning to homepage in future versions.