Add json data files, add new API call, create reservation page prototype
Jsons are loaded into database when tables are created (if not existing). New API added for checking user's reservations. Known bug: actually database creating is buggy because of foreign keys creating errors when inserting data, javascript insert data asynchronously in the database but it should wait for other tables to create and insert data to avoid this constraints.