Add carousel items dynamic population in utils.js, add images to artists
Now carousel items and indicators are generated automatically from a list of image paths and appended to respective containers. This functions can be used in every page with a photo gallery. Image paths are fetched from a database column of type array, this type of column should be inserted to every table objects with a photo gallery (for now only in artists, need to update data in json).